pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-quality-lab"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9",
    "openai>=1.30",
    "tiktoken>=0.7",
    "rank_bm25>=0.2.2",
    "numpy>=1.26",
    "pandas>=2.2",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "rich>=13.7",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-mock>=3.14",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["rag_quality_lab*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: requires docker compose up"]
